Renamed the AAP authentication provider and upped the priority
